add more verbosity and update test_qp/main_test.cpp to test the CGAL quadratic solver from iterators

+++ includes/cpr_graphmatching_cgal.h
@@ -37,7 +37,7 @@ public:
   GraphMatchingCgal(Eigen::MatrixXd const *vsim, EdgeSimilarityMatrix const *esim, Eigen::MatrixXi const *g_adj, Eigen::MatrixXi const *h_adj);
   virtual void run();
   void fillMpsStream(std::stringstream &in) const;
-  void fillQuadraticObjective(std::vector<int *> &D);       // objective, rowwise, on/below diagonal only, multiplied by a factor 2
+  void fillQuadraticObjective(std::vector<double *> &D);    // objective, rowwise, on/below diagonal only, multiplied by a factor 2
   void fillStochasticityConstraints(std::vector<int *> &A); // make constraints, columnwise
   //unsigned int mappedVertex(unsigned int) const;
 };
